refactor: getCatById, mapping returned object for the correct layer

+++ src/main/java/org/puzre/core/port/service/ICatService.java
@@ -13,7 +13,7 @@ public interface ICatService {
     Page<Cat> listAllCats(Integer page, Integer totalItems);
     List<Cat> listCatsLegacyByType(int typeId);
     PaginatedResponse<CatEntity, Cat> listCatsByType(int typeId, int page, int totalItems);
-    Cat findCatById(int catId);
+    Cat findCatById(Long catId);
     List<Cat> searchCatsByMessageLegacy(String message);
     PaginatedResponse<CatEntity, Cat> searchCatsByMessage(String message, int page, int totalItems);
